# swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Impera Backend API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/games/open": {
      "get": {
        "tags": [
          "Game"
        ],
        "summary": "Get a list of open games, excluding games by the current player",
        "operationId": "Game_GetAll",
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "List of games",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/games/my": {
      "get": {
        "tags": [
          "Game"
        ],
        "summary": "Get a list of the games for the current player",
        "operationId": "Game_GetMy",
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "List of games for the current user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/games/myturn": {
      "get": {
        "tags": [
          "Game"
        ],
        "summary": "Get list of games where it's the current player's team",
        "operationId": "Game_GetMyTurn",
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "List of games where it's the current user's team",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/games": {
      "post": {
        "tags": [
          "Game"
        ],
        "summary": "Create a new game",
        "operationId": "Game_Post",
        "requestBody": {
          "x-name": "creationOptions",
          "description": "Creation options",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameCreationOptions"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Summary of newly created game",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameSummary"
                }
              }
            }
          }
        }
      }
    },
    "/games/{gameId}": {
      "get": {
        "tags": [
          "Game"
        ],
        "summary": "Get detailed information about a single game",
        "operationId": "Game_Get",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "description": "Id of the requested game",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Information about the requested game",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Game"
        ],
        "summary": "Cancel/delete the requested game, if possible.",
        "description": "This is only possible, if the requested game is in a state that\ncan be deleted",
        "operationId": "Game_Delete",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "description": "Id of the game to delete",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Status"
          }
        }
      }
    },
...
    "/games/{gameId}/hide": {
      "patch": {
        "tags": [
          "Game"
        ],
        "summary": "Hides the given game for the current player",
        "operationId": "Game_PatchHide",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "description": "Id of game to hide",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": ""
          }
        }
      }
    },
    "/games/hide": {
      "patch": {
        "tags": [
          "Game"
        ],
        "summary": "Hide all games which can be hidden for the current player",
        "operationId": "Game_PatchHideAll",
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "error_Description": {
            "type": "string",
            "nullable": true
          },
          "parameter_Errors": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "GameSummary": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "$ref": "#/components/schemas/GameType"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "hasPassword": {
            "type": "boolean"
          },
          "ladderId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "ladderName": {
            "type": "string",
            "nullable": true
          },
          "options": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/GameOptions"
              }
            ]
          },
          "createdByUserId": {
            "type": "string",
            "nullable": true
          },
          "createdByName": {
            "type": "string",
            "nullable": true
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastActionAt": {
            "type": "string",
            "format": "date-time"
          },
          "timeoutSecondsLeft": {
            "type": "integer",
            "format": "int32"
          },
          "mapTemplate": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "$ref": "#/components/schemas/GameState"
          },
          "currentPlayer": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/PlayerSummary"
              }
            ]
          },
          "teams": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TeamSummary"
            }
          },
          "turnCounter": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "GameType": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Fun",
          "Ranking",
          "Tournament"
        ],
        "enum": [
          "Fun",
          "Ranking",
          "Tournament"
        ]
      },
      "GameOptions": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "numberOfPlayersPerTeam",
          "numberOfTeams",
          "attacksPerTurn",
          "movesPerTurn",
          "mapDistribution",
          "timeoutInSeconds",
          "maximumNumberOfCards",
          "victoryConditions",
          "visibilityModifier"
        ],
        "properties": {
          "numberOfPlayersPerTeam": {
            "type": "integer",
            "format": "int32",
            "maximum": 8.0,
            "minimum": 1.0
          },
          "numberOfTeams": {
            "type": "integer",
            "format": "int32",
            "maximum": 16.0,
            "minimum": 2.0
          },
          "minUnitsPerCountry": {
            "type": "integer",
            "format": "int32",
            "maximum": 5.0,
            "minimum": 0.0
          },
          "newUnitsPerTurn": {
            "type": "integer",
            "format": "int32",
            "maximum": 10.0,
            "minimum": 3.0
          },
          "attacksPerTurn": {
            "type": "integer",
            "format": "int32",
            "maximum": 100.0,
            "minimum": 1.0
          },
          "movesPerTurn": {
            "type": "integer",
            "format": "int32",
            "maximum": 100.0,
            "minimum": 1.0
          },
          "initialCountryUnits": {
            "type": "integer",
            "format": "int32",
            "maximum": 5.0,
            "minimum": 0.0
          },
          "mapDistribution": {
            "$ref": "#/components/schemas/MapDistribution"
          },
          "timeoutInSeconds": {
            "type": "integer",
            "format": "int32",
            "maximum": 432000.0,
            "minimum": 180.0
          },
          "maximumTimeoutsPerPlayer": {
            "type": "integer",
            "format": "int32"
          },
          "maximumNumberOfCards": {
            "type": "integer",
            "format": "int32",
            "maximum": 10.0,
            "minimum": 0.0
          },
          "victoryConditions": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/VictoryConditionType"
            }
          },
          "visibilityModifier": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/VisibilityModifierType"
            }
          }
        }
      },
      "MapDistribution": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Default",
          "Malibu",
          "Malibu3"
        ],
        "enum": [
          "Default",
          "Malibu",
          "Malibu3"
        ]
      },
      "VictoryConditionType": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Survival",
          "ControlContinent",
          "Capitals",
          "Rush"
        ],
        "enum": [
          "Survival",
          "ControlContinent",
          "Capitals",
          "Rush"
        ]
      },
      "VisibilityModifierType": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "None",
          "Fog"
        ],
        "enum": [
          "None",
          "Fog"
        ]
      },
      "GameState": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "None",
          "Open",
          "Active",
          "Ended"
        ],
        "enum": [
          "None",
          "Open",
          "Active",
          "Ended"
        ]
      },
      "PlayerSummary": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "$ref": "#/components/schemas/PlayerState"
          },
          "outcome": {
            "$ref": "#/components/schemas/PlayerOutcome"
          },
          "teamId": {
            "type": "string",
            "format": "guid"
          },
          "playOrder": {
            "type": "integer",
            "format": "int32"
          },
          "timeouts": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PlayerState": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "None",
          "Active",
          "InActive"
        ],
        "enum": [
          "None",
          "Active",
          "InActive"
        ]
      },
      "PlayerOutcome": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "None",
          "Won",
          "Defeated",
          "Surrendered",
          "Timeout"
        ],
        "enum": [
          "None",
          "Won",
          "Defeated",
          "Surrendered",
          "Timeout"
        ]
      },
      "TeamSummary": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "playOrder": {
            "type": "integer",
            "format": "int32"
          },
          "players": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/PlayerSummary"
            }
          }
        }
      },
      "GameCreationOptions": {
        "allOf": [
          {
            "$ref": "#/components/schemas/GameOptions"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "name",
              "mapTemplate"
            ],
            "properties": {
              "name": {
                "type": "string",
                "maxLength": 64,
                "minLength": 3
              },
              "password": {
                "type": "string",
                "maxLength": 64,
                "minLength": 3,
                "nullable": true
              },
              "addBot": {
                "type": "boolean"
              },
              "mapTemplate": {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              }
            }
          }
        ]
      },
      "Game": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "$ref": "#/components/schemas/GameType"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "hasPassword": {
            "type": "boolean"
          },
          "mapTemplate": {
            "type": "string",
            "nullable": true
          },
          "teams": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Team"
            }
          },
          "state": {
            "$ref": "#/components/schemas/GameState"
          },
          "playState": {
            "$ref": "#/components/schemas/PlayState"
          },
          "currentPlayer": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/PlayerSummary"
              }
            ]
          },
          "map": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Map"
              }
            ]
          },
          "options": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/GameOptions"
              }
            ]
          },
          "lastModifiedAt": {
            "type": "string",
            "format": "date-time"
          },
          "timeoutSecondsLeft": {
            "type": "integer",
            "format": "int32"
          },
          "turnCounter": {
            "type": "integer",
            "format": "int32"
          },
          "unitsToPlace": {
            "type": "integer",
            "format": "int32"
          },
          "attacksInCurrentTurn": {
            "type": "integer",
            "format": "int32"
          },
          "movesInCurrentTurn": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Team": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "playOrder": {
            "type": "integer",
            "format": "int32"
          },
          "players": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Player"
            }
          }
        }
      },
      "Player": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PlayerSummary"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "cards": {
                "type": "array",
                "nullable": true,
                "items": {
                  "$ref": "#/components/schemas/BonusCard"
                }
              },
              "placedInitialUnits": {
                "type": "boolean"
              },
              "numberOfUnits": {
                "type": "integer",
                "format": "int32"
              },
              "numberOfCountries": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ]
      },
      "BonusCard": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "A",
          "B",
          "C"
        ],
        "enum": [
          "A",
          "B",
          "C"
        ]
      },
      "PlayState": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "None",
          "PlaceUnits",
          "Attack",
          "Move",
          "Done"
        ],
        "enum": [
          "None",
          "PlaceUnits",
          "Attack",
          "Move",
          "Done"
        ]
      },
      "Map": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "countries": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Country"
            }
          }
        }
      },
      "Country": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "identifier": {
            "type": "string",
            "nullable": true
          },
          "playerId": {
            "type": "string",
            "format": "guid"
          },
          "teamId": {
            "type": "string",
            "format": "guid"
          },
          "units": {
            "type": "integer",
            "format": "int32"
          },
          "flags": {
            "$ref": "#/components/schemas/CountryFlags"
          }
        }
      },
      "CountryFlags": {
        "type": "string",
        "description": "",
        "x-enumFlags": true,
        "x-enumNames": [
          "None",
          "Capital"
        ],
        "enum": [
          "None",
          "Capital"
        ]
      },
...
    }
  }
}